refactor: separated LoginForm from LoginPage
